Give mock.Arguments a method to assign arguments to named return values.
To return a nill pointer via mock.Called().Get().(*type) requires a two-value type assertion and a conditional because nil fails type assertion to pointer types.

Instead, passing named return values to a method by reference means we can avoid using type assertion. This also shrinks the mock method a bit as the user doesn't have to write out a line of type assertions.

Further, we can give more useful error messages to the user, such as when more arguments are passed to Return() than are unpacked with Get(), which is a silent fault today.